Various fixes to get graphene to display my dash
fixed a bug(?) in the GaugeGadget where it wouldn't display anything if
data[] only had one value in it.
TimeSeries legend will show the current value in yellow by default along
with the min and the max
added a 'show' parameter which allows selectively showing the min/max/cur
values in the legend.
added a 'start_color' parameter which allows changing the color which
graphene will use to show the first plot
added a 'value_format' option which allows setting a custom formatter
for the legend min/max/cur values