Aggregate Series chart type and various bug fixes
increased title font in css to 1.5em.

Allow a custom model to be used using a new model: {} option in the js.

fixed a possible bug where the min or max value of a set was 0, if it
was then graphene would just throw the whole set away.

new AggregateSeries model which takes an initial source URL, passes the
targets through some custom aggregation functions and then will make
another request for the desired graph targets. This is useful in
displaying various things about a 'top x' sample.

added a 'noarea' option to the view parameters, which if set to true will
not draw the area under the line

added a 'nosort' option to the view parameters, which if set to true
will not do any sorting of the data series. Useful for Aggregate series
if you want to have many graphs with the same ordering independent of
their values.

added a 'logscale' option to the view parameters, which if set to true
will display a logarithmic scale for the y axis.

fixed a bug with start_color which would be unset after the first
gathering of data sometimes